move debugging methods to packets.cpp to avoid circular dependancy. packets depends on util_packets
